Feeing the rabbit a carrot
